requirements.txt
pillow==6.2.1

